+++ backend/tests/Designer.Tests/Controllers/PreviewController/AllowedToInstantiateFilterTests.cs
@@ -1,0 +1,7 @@
+﻿namespace Designer.Tests.Controllers.PreviewController
+{
+    public class AllowedToInstantiateFilterTests
+    {
+        
+    }
+}
