+++ backend/tests/Designer.Tests/Controllers/PreviewController/KeepAliveTests.cs
@@ -1,0 +1,26 @@
+﻿using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Xunit;
+
+namespace Designer.Tests.Controllers.PreviewController
+{
+    public class KeepAliveTests : PreviewControllerTestsBase<KeepAliveTests>
+    {
+
+        public KeepAliveTests(WebApplicationFactory<Altinn.Studio.Designer.Controllers.PreviewController> factory) : base(factory)
+        {
+        }
+
+        [Fact]
+        public async Task Get_KeepAlive_Ok()
+        {
+            string dataPathWithData = $"{Org}/{App}/api/authentication/keepAlive";
+            using HttpRequestMessage httpRequestMessage = new(HttpMethod.Get, dataPathWithData);
+
+            using HttpResponseMessage response = await HttpClient.Value.SendAsync(httpRequestMessage);
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+        }
+    }
+}
