+++ backend/tests/Designer.Tests/Controllers/PreviewController/LayoutSetsTests.cs
@@ -1,0 +1,25 @@
+﻿using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Xunit;
+
+namespace Designer.Tests.Controllers.PreviewController
+{
+    public class LayoutSetsTests : PreviewControllerTestsBase<LayoutSetsTests>
+    {
+        public LayoutSetsTests(WebApplicationFactory<Altinn.Studio.Designer.Controllers.PreviewController> factory) : base(factory)
+        {
+        }
+
+        [Fact]
+        public async Task Get_LayoutSets_NotFound()
+        {
+            string dataPathWithData = $"{Org}/{App}/api/layoutsets";
+            using HttpRequestMessage httpRequestMessage = new(HttpMethod.Get, dataPathWithData);
+
+            using HttpResponseMessage response = await HttpClient.Value.SendAsync(httpRequestMessage);
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+        }
+    }
+}
